% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{Fonction pour avoir le nombre de lignes d'un dataframe}
\usage{
get_df_nrows(df)
}
\arguments{
\item{df}{: dataframe to use}
}
\value{
nrow_df
}
\description{
Description
Get a message with a number of rows in a data frame.
}
\examples{
df <- iris
get_df_nrows(df)
df <- mtcars
get_df_nrows(df)

}
